Add notes. Move replace.
