+++ Logger.py
@@ -3,6 +3,7 @@ from datetime import datetime
 
 
 def Logger(database_folder):
+  """Save log file to Serato database"""
   log_filename = f'{os.path.splitext(os.path.basename(__file__))[0]}-{datetime.now().strftime("%Y-%m-%d")}.log'
   log = f'{database_folder}/Logs/{log_filename}'
   logging.basicConfig(filename=log, level=logging.DEBUG, format='%(asctime)s %(levelname)s %(message)s', datefmt='%Y-%m-%d %H:%M:%S', force=True)
